requirements.txt
aiohttp==3.7.4.post0
asttokens==2.0.5
async-timeout==3.0.1
attrs==21.4.0
autopep8==1.6.0
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
colorama==0.4.5
Deprecated==1.2.13
discord.py==1.7.3
executing==0.9.1
gitdb==4.0.9
GitPython==3.1.27
i==2.1.0
idna==3.3
multidict==6.0.2
pep8==1.7.1
pycodestyle==2.8.0
pycparser==2.21
pygit2==1.10.0
PyGithub==1.55
Pygments==2.12.0
PyJWT==2.4.0
PyNaCl==1.5.0
python-decouple==3.6
requests==2.28.1
six==1.16.0
smmap==5.0.0
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.11
wrapt==1.14.1
yarl==1.7.2
